requirements.txt
django>=3.0.6,<3.1
uWSGI>=2.0.18,<2.1
psycopg2-binary~=2.8.6
fpdf
boto3